agregamos token y carrousel

+++ backend/controllers/inmueble/deleteRelacionUserInmueble.js
@@ -1,7 +1,18 @@
-const pool = require("../../config/db")
+const pool = require("../../config/db");
+const { verifyToken } = require("../../middlewares/jwt");
 
 const deleteRelacionUserInmueble = async(req, res)=>{
     try {
+        const token = req.cookies.token;
+        console.log(token)
+        if (!token) {
+            return res.status(401).json({
+                status: false,
+                message: "Sesion no iniciada, inicie de nuevo.",
+                error: "Token no encontrado."
+            });
+        }
+        verifyToken(token)
         const {idUser, idInmueble} = req.body
         const queryDelete = "delete from user_inmueble where id_user = $1 and id_inmueble = $2"
         const query = await pool.query(queryDelete, [idUser, idInmueble])
